requirements.txt
anyio==3.5.0
asgiref==3.4.1
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.71.0
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-language==2.3.1
google-cloud-texttospeech==2.9.0
google-cloud-vision==2.6.3
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.12.0
httplib2==0.20.2
httptools==0.2.0
idna==3.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
motor==3.0.0
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.5.62
orjson==3.6.5
praw==7.5.0
prawcore==2.3.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.0
pydub==0.25.1
pymongo==4.1.1
pyparsing==3.0.6
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
ujson==5.1.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websocket-client==1.2.3
websockets==10.1
